+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 -   `Host::getAll()`/`Host::get()`: Fetch all "effective" attributes from hosts by default
+-   `Folder::get()`/`Folder::getAll()`: Switch from `output_format=python` to `output_format=json`
 
 
 ### Fixed